requirements.txt
torch
PyGithub
requests
chromadb
sentence-transformers
python-dotenv
accelerate